# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Taskts",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "This is Task app"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    },
    {
      "url": "https://timemanagement-back.onrender.com"
    }
  ],
...
  "paths": {
...
    "/note": {
      "get": {
        "tags": [
          "Notes"
        ],
        "summary": "Get all Notes",
        "operationId": "getNotesController",
        "description": "Get list of users with this endpoint",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            },
            "example": 1
          },
          {
            "in": "query",
            "name": "perPage",
            "schema": {
              "type": "integer",
              "example": 10
            }
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string",
              "example": "sortOrder"
            },
            "description": "All Note's fields. Default value - _id"
          },
          {
            "in": "query",
            "name": "sortOrder",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ],
              "example": "asc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully found Notes!",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully found Notes!"
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "data",
                        "page",
                        "perPage",
                        "totalItems",
                        "totalPages",
                        "hasNextPage",
                        "hasPreviousPage"
                      ],
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/notes"
                          }
                        },
                        "page": {
                          "type": "number",
                          "example": 2
                        },
                        "perPage": {
                          "type": "number",
                          "example": 4
                        },
                        "totalItems": {
                          "type": "number",
                          "example": 4
                        },
                        "totalPages": {
                          "type": "number",
                          "example": 4
                        },
                        "hasNextPage": {
                          "type": "boolean",
                          "example": false
                        },
                        "hasPreviousPage": {
                          "type": "boolean",
                          "example": true
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      },
      "post": {
        "tags": [
          "Notes"
        ],
        "summary": "Create note",
        "operationId": "createNote",
        "description": "Create a note with payload",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "note"
                ],
                "properties": {
                  "note": {
                    "definitions": "note name",
                    "type": "string",
                    "example": "dddhh aayyz"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Creates a note",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 201
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully created a note!"
                    },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/notes"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      }
    },
    "/note/{id}": {
      "get": {
        "tags": [
          "Notes"
        ],
        "summary": "Get Note by id",
        "operationId": "getNoteIDController",
        "description": "Get Note by id",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "example": "6692528c222b4be7b4b17ebd"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get Note by id",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully found Note with id 6692528c222b4be7b4b17ebd!"
                    },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/notes"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      },
      "delete": {
        "noteRouterNotestags": [
          "Notes"
        ],
        "summary": "Delete note",
        "operationId": "deleteNote",
        "description": "Delete a note from DB",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "example": "6692528c222b4be7b4b17ebd"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Delete note by id"
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      },
      "patch": {
        "tags": [
          "Notes"
        ],
        "summary": "Change Note by ID",
        "operationId": "ChangeNote",
        "description": "Change a Note by ID",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "example": "6692528c222b4be7b4b17ebd"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "note": {
                    "definitions": "Note name",
                    "type": "string",
                    "example": "dddhh aayyz"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Change a Note",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully changed a Note!"
                    },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/notes"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "notes": {
        "type": "object",
        "required": [
          "_id",
          "note",
          "userId"
        ],
        "properties": {
          "_id": {
            "definitions": "note's ID",
            "type": "string",
            "example": "6692528c222b4be7b4b17ebd"
          },
          "note": {
            "definitions": "note",
            "type": "string",
            "example": "dddhh aayyz"
          },
          "timeDeclaration": {
            "definitions": "120",
            "type": "number",
            "example": 120
          },
          "timeReal": {
            "definitions": "120",
            "type": "number",
            "example": 120
          },
          "isFavourite": {
            "definitions": "Would you like or not?",
            "type": "boolean",
            "example": false
          },
          "userId": {
            "definitions": "Autorth's note ID",
            "type": "string",
            "example": "6692528c222b4be7b4b17ebd"
          }
        }
      },
...
    },
    "responses": {
...
      "404": {
        "description": "Unauthorized error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 404
                },
                "message": {
                  "type": "string",
                  "example": "Route not found"
                },
                "data": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Route not found"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}